# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "description": "The error envelope every non-2xx response carries.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail",
            "description": "The error itself."
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
      "ErrorDetail": {
        "description": "Machine-readable error code plus a human-readable message.",
        "properties": {
          "code": {
            "description": "Stable machine-readable code (e.g. `invalid_cursor`, `not_found`).",
            "examples": [
              "invalid_cursor"
            ],
            "type": "string"
          },
          "message": {
            "description": "Human-readable explanation.",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Worldwide politician financial-disclosure tracking. Cursor pagination on ULIDs; consistent error envelope; verification_state on every record. Every successful GET carries a strong ETag (sha256 of the body); requests with a matching If-None-Match receive 304 Not Modified. Authentication: `Authorization: Bearer gfk_...` API keys (create via /v1/keys; hashes only at rest). Requests without a key are served at the free tier. Tiers control record freshness (free: filings we discovered less than 24 hours ago are not yet visible; pro/data: real time) and the daily request quota (counted per account; over quota is 429 quota_exceeded in the standard envelope). Anonymous traffic has a light per-IP rate backstop (429 rate_limited).",
    "title": "govfolio API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/stripe/webhook": {
      "post": {
        "description": "# Errors\n`503` when no webhook secret is configured; `400` on a missing/invalid\nsignature or an unparseable event; `500` on backend failure.",
        "operationId": "stripe_webhook",
        "requestBody": {
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              }
            }
          },
          "description": "Raw Stripe event JSON (signature over the exact bytes)",
          "required": true
        },
        "responses": {
          "200": {
            "description": "Event verified and processed (or acknowledged as not ours)"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing/invalid Stripe-Signature or malformed event"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Webhook secret not configured (fail closed)"
          }
        },
        "summary": "Receives Stripe webhook events (subscription lifecycle mirror).",
        "tags": [
          "billing"
        ]
      }
    },
...
  },
...
}